requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
autogluon.tabular
fastapi
uvicorn
joblib
requests
openaq
loguru
